Added 'img' method to PaHaWDataset
This allows users to work with 2d image arrays of the handwriting data.
Beware that the PaHaWDataset.x attribute in this mode consists of a
tuple of 3d image arrays with shape [y, x, image]. The first consists of
the on-paper image, and the second consists of the in-air image.